+++ src/comments/comment.css
@@ -50,6 +50,7 @@
   }
   
   .data {
+    display: -ms-grid;
     display: grid;
     grid-template-columns: auto auto;
     justify-content: space-between;
